+++ src/callbacks/preGetCollectible.ts
@@ -1,5 +1,8 @@
+import { log } from "isaacscript-common";
 import * as reverseJusticeFix from "../features/optional/bugfix/reverseJusticeFix";
 import { betterDevilAngelRoomsPreGetCollectible } from "../features/optional/major/betterDevilAngelRooms/callbacks/preGetCollectible";
+
+const DEBUG = false;
 
 export function init(mod: Mod): void {
   mod.AddCallback(ModCallbacks.MC_PRE_GET_COLLECTIBLE, main);
@@ -10,11 +13,11 @@ function main(
   decrease: boolean,
   seed: int,
 ): CollectibleType | int | void {
-  /*
-  log(
-    `MC_PRE_GET_COLLECTIBLE - itemPoolType: ${itemPoolType}, decrease: ${decrease}, seed: ${seed}`,
-  );
-  */
+  if (DEBUG) {
+    log(
+      `MC_PRE_GET_COLLECTIBLE - itemPoolType: ${itemPoolType}, decrease: ${decrease}, seed: ${seed}`,
+    );
+  }
 
   // This has to be before the "betterDevilAngelRooms" feature
   const returnValue = reverseJusticeFix.preGetCollectible();